feature: migrate indexeddb to store/browser.js, export store for node, update imports ternary, use indexeddb in SwapCreate, SwapActivity/ActivityItem, fixes for network/browser.js
