+++ js/sdk/lib/store/index.js
@@ -7,6 +7,8 @@ const isBrowser = new Function('try {return this===window;}catch(e){ return fals
 // const isNode = new Function('try {return this===global;}catch(e){return false;}')
 /* eslint-enable no-new-func */
 
-module.exports = isBrowser()
-  ? require('./browser')
-  : require('./node')
+const { Store } = isBrowser()
+? require('./browser')
+: require('./node');
+
+module.exports = Store;
