+++ js/sdk/lib/store/node.js
@@ -14,7 +14,7 @@ const INSTANCES = new WeakMap()
  * Store implementation for the node.js
  * @type {Store}
  */
-module.exports = class Store extends BaseClass {
+class Store extends BaseClass {
   constructor (props) {
     super()
 
@@ -136,3 +136,5 @@ module.exports = class Store extends BaseClass {
     return Promise.resolve()
   }
 }
+
+module.exports = { Store };
